1.Now we can review histories of Blig status. 2.Bugs fixed on Blig status. 3."Run" button added.
In this program, we check the resources that each action need backward. We will do this action if we have enough resources. Then we can optimize it by the heuristic function.
For example, the Plinks will continue find plonks even we have 1000 plonks already, which should be optimized.
Try this program. Feel free to let me know if you find any bugs. :)