Added a new suggestion-calculating algorithm in a separate project. The algorithm uses a graph representation with our adjustable resources as nodes.
The new project implements the algorithm in a new Form that launches along with the main Form.

When testing it has consistently given suggestions that improve the plan, converging on 29 hours. If we try to run the algorithm on the excess of a 29h long plan, the generated suggestion will result in a plan that's roughly 36h long, and will then start converging on 29h again.

-[The algorithm]---------------------------------

Each node is annotated with the average amount of excess units to redistributed. Additionally each node is annotated with the suggested adjustment to that resource type (initialized to 0).

Any two resources that appear together in the precondition of at least 1 action are connected in the graph. Each connection has 2 values: the amount that can be through that link (called "from" value) and the resulting amount of adjustment to the receiving node (called the "to" value).

Excess resources can be moved from a given node if, and only if, there is enough excess left in the node to distribute uniformly across all connections to that node. This will result in a negative change to the adjustment value of the node, and a positive change to the adjustment value of it's neighbors.

When a the adjustment value of a node is increased, the excess value of the neighboring nodes should be decreased by a corresponding amount. This reflects that we can now potentially perform more actions using the previously unused resources.

The algorithm halts when we can no longer move any excess resources.